Bug 187 fixconflict (#5)
* (m, PEP8) Fix long lines

* Fix bug-269: store Autoscroll tool state in config

TaurusTrend2D does not "remember" the state of its AutoScroll setting.
Fix by making the AutoScrollTool object inherit from
BaseConfigurableClass, and registering its check state (and registering
the tool as a config delegate for the TaurusTrend2DDialog).

Also: register the tool as a config delegate for TaurusTrend1DDialog too

* Fix scale problem when autoscroll is set before model

In TaurusTrend2D, setting the autoscroll before the model triggers a bug
that permanently sets the scale range to min,max = 0,0, preventing
correct display of the data. Fix it by ignoring scroll requests before
the axis is properly set.